requirements.txt
# Python deps for `build_site.py` (blog from Supabase + dotenv). Used by Vercel build and local `pip install -r requirements.txt`.
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
python-dotenv>=1.0.0
